Split Operations into two files
